start implementing ansi network support
