+++ CMakeLists.txt
@@ -131,6 +131,7 @@ check_include_files(windows.h HAVE_WINDOWS_H)
 check_include_files(winsock2.h HAVE_WINSOCK2_H)
 
 check_symbol_exists(fopen_s stdio.h HAVE_FOPEN_S)
+check_symbol_exists(getaddrinfo netdb.h HAVE_GETADDRINFO)
 check_symbol_exists(gethostname unistd.h HAVE_UNISTD_GETHOSTNAME)
 check_symbol_exists(getpagesize unistd.h HAVE_UNISTD_GETPAGESIZE)
 check_symbol_exists(gmtime time.h HAVE_GMTIME)
@@ -267,13 +268,17 @@ if(HAVE_FOPEN_S)
   list(APPEND STUMPLESS_SOURCES src/config/have_fopen_s.c)
 endif(HAVE_FOPEN_S)
 
-if(HAVE_GMTIME)
-  list(APPEND STUMPLESS_SOURCES src/config/have_gmtime.c)
+if(HAVE_GETADDRINFO)
+  list(APPEND STUMPLESS_SOURCES src/config/have_getaddrinfo.c)
+else()
+  list(APPEND STUMPLESS_SOURCES src/config/no_getaddrinfo.c)
 endif()
 
 if(HAVE_GMTIME_R)
   list(APPEND STUMPLESS_SOURCES src/config/have_gmtime_r.c)
-endif(HAVE_GMTIME_R)
+elseif(HAVE_GMTIME)
+  list(APPEND STUMPLESS_SOURCES src/config/have_gmtime.c)
+endif()
 
 if(HAVE_PTHREAD_H)
   list(APPEND STUMPLESS_SOURCES src/config/have_pthread.c)
