+++ include/private/config.h.in
@@ -31,6 +31,7 @@
 
 /* symbol checks */
 #cmakedefine HAVE_FOPEN_S 1
+#cmakedefine HAVE_GETADDRINFO 1
 #cmakedefine HAVE_GMTIME 1
 #cmakedefine HAVE_GMTIME_R 1
 #cmakedefine HAVE_UNISTD_SC_PAGESIZE 1
