+++ tools/check_headers/stumpless_private.yml
@@ -24,6 +24,8 @@
 "config_init_journald_element": "private/config/wrapper/journald.h"
 "config_init_journald_param": "private/config/wrapper/journald.h"
 "config_init_mutex": "private/config/wrapper/thread_safety.h"
+"config_initialize_wel_data": "private/config/wrapper/wel.h"
+"config_int_connect": "private/config/wrapper/int_connect.h"
 "config_journald_free_thread": "private/config/wrapper/journald.h"
 "config_lock_mutex": "private/config/wrapper/thread_safety.h"
 "config_read_flag": "private/config/wrapper/thread_safety.h"
@@ -33,7 +35,6 @@
 "config_close_wel_target": "private/config/wrapper/wel.h"
 "config_copy_wel_data": "private/config/wrapper/wel.h"
 "config_destroy_wel_data": "private/config/wrapper/wel.h"
-"config_initialize_wel_data": "private/config/wrapper/wel.h"
 "config_sendto_wel_target": "private/config/wrapper/wel.h"
 "config_set_entry_wel_type": "private/config/wrapper/wel.h"
 "CONFIG_THREAD_LOCAL_STORAGE": "private/config/wrapper/thread_safety.h"
@@ -55,8 +56,11 @@
 "get_event_id": "private/config/wel_supported.h"
 "get_journald_field_name": "private/target/journald.h"
 "get_type": "private/config/wel_supported.h"
+"getaddrinfo_int_connect": "private/config/have_getaddrinfo.h"
 "gmtime_get_now": "private/config/have_gmtime.h"
 "HAVE_FOPEN_S": "private/config.h"
+"HAVE_GETADDRINFO": "private/config.h"
+"HAVE_GMTIME": "private/config.h"
 "HAVE_PTHREAD_H": "private/config.h"
 "HAVE_STRNCPY_S": "private/config.h"
 "HAVE_VSNPRINTF_S": "private/config.h"
@@ -88,6 +92,7 @@
 "locked_swap_wel_insertion_string": "private/config/wel_supported.h"
 "new_entry": "private/entry.h"
 "no_abstract_socket_names_get_local_socket_name": "private/config/abstract_socket_names_unsupported.h"
+"no_getaddrinfo_int_connect": "private/config/no_getaddrinfo.h"
 "no_thread_safety_compare_exchange_bool": "private/config/thread_safety_unsupported.h"
 "no_thread_safety_compare_exchange_ptr": "private/config/thread_safety_unsupported.h"
 "no_wcsrtombs_s_copy_wstring_to_cstring": "private/config/no_wcsrtombs_s.h"
